requirements.txt
# Python dependencies for resume PDF/LaTeX generation
weasyprint==68.1
Jinja2==3.1.6
PyYAML==6.0.3
